fix index out of range bug in u_UCharToChars
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/336



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@974 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/putil.c
@@ -1448,7 +1448,7 @@ u_UCharsToChars(const UChar *us, char *cs, UTextOffset length) {
 #if U_CHARSET_FAMILY==U_ASCII_FAMILY
         *cs++=(char)(*us++);
 #elif U_CHARSET_FAMILY==U_EBCDIC_FAMILY
-        *cs++=(char)ebcdicFromAscii[*us++];
+        *cs++=(char)ebcdicFromAscii[(uint8_t)(*us++)];
 #else
 #   error U_CHARSET_FAMILY is not valid
 #endif
